Added a bunch more plugins, fixed a number of config issues, and polished.
This represents a fair amount of work to integrate some of the things a
co-worker showed me recently--my thanks to you, Alex G.! I wouldn't
suggest using this as an example of how to do things right. I'm still
puzzled over making the hydra plugin work, and my first take on
switching from lualine to heirline for a status bar is practically a
graduate level course in how NOT to write code. But I do think a number
of things have been improved here nevertheless.